api/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flask-cors==5.0.1
python-dotenv==0.19.0
requests==2.32.3

gunicorn
psycopg==3.2.9
pytz==2024.1
numpy==2.3.1
pandas==2.3.0
longport==3.0.7
